Move check_status_code back to utils; move tests
