+++ tests/test_upload.py
@@ -298,22 +298,7 @@ def test_check_status_code_for_wrong_repo_url(repo_url, make_settings):
     # override defaults to use incorrect URL
     upload_settings.repository_config['repository'] = repo_url
 
-    with pytest.raises(exceptions.InvalidPyPIUploadURL):
+    with pytest.raises(twine.exceptions.InvalidPyPIUploadURL):
         upload.upload(upload_settings, [
             WHEEL_FIXTURE, SDIST_FIXTURE, NEW_SDIST_FIXTURE, NEW_WHEEL_FIXTURE
         ])
-
-
-@pytest.mark.parametrize('repo_url', [
-    "https://pypi.python.org",
-    "https://testpypi.python.org"
-])
-def test_check_status_code_for_deprecated_pypi_url(repo_url):
-    response = pretend.stub(
-        status_code=410,
-        url=repo_url
-    )
-
-    # value of Verbose doesn't matter for this check
-    with pytest.raises(exceptions.UploadToDeprecatedPyPIDetected):
-        upload.check_status_code(response, False)
